pyproject.toml
[project]
name = "techideias-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nicegui>=3.3.1",
    "python-dotenv>=1.2.1",
    "secure-smtplib>=0.1.1",
]
